Rename storage/v1 package and move migrations
