+++ storage/psql/config.go
@@ -1,4 +1,4 @@
-package v1
+package psql
 
 import (
 	"flag"
